+++ app/src/main/java/no/kristiania/myapplication/models/CryptoCurrencies.kt
@@ -6,8 +6,8 @@ import java.io.Serializable
 data class CryptoCurrencies(
     var symbol:String,
     var name:String,
-    var priceUsd: Float,
-    var changePercent24Hr: Float,
+    var priceUsd: Int,
+    var changePercent24Hr: Int,
     var id:String) : Serializable
